fix(orchestration): use typed DB methods for inbox and tighten coordinator scope
